+++ CHANGELOG.md
@@ -11,3 +11,6 @@
 
 - ghcr.io/prymitive/karma is now used as the primary repository for docker
   images instead of Docker Hub
+- `alertAcknowledgement:commentPrefix` config option was replaced by
+  `alertAcknowledgement:comment` that can be used to customise the entire
+  comment.